+++ src/test/scala/edu/ie3/simona/model/thermal/ThermalGridTestData.scala
@@ -25,4 +25,8 @@ trait ThermalGridTestData {
   protected val testGridQDotInfeed: Power = Kilowatts(15d)
   protected val testGridQDotConsumption: Power = Kilowatts(-42d)
   protected val testGridQDotConsumptionHigh: Power = Kilowatts(-200d)
+  protected val noThermalDemand: Boolean = false
+  protected val thermalDemand: Boolean = true
+  protected val isRunning: Boolean = true
+  protected val isNotRunning: Boolean = false
 }